# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sineas Video Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Video {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  uploaderName: string;
  uploaderId?: string;
  videoUrl: string;
  /** @nullable */
  url_360p?: string | null;
  /** @nullable */
  url_480p?: string | null;
  /** @nullable */
  url_720p?: string | null;
  /** @nullable */
  url_1080p?: string | null;
  /** @nullable */
  url_4k?: string | null;
  /** @nullable */
  thumbnailUrl: string | null;
  /** Duration in seconds */
  duration: number;
  viewCount: number;
  likeCount: number;
  /** @nullable */
  genre?: string | null;
  isPublic: boolean;
  isPremium: boolean;
  /**
     * basic, premium, or ultra
     * @nullable
     */
  minimumPlan?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface VideoList {
  videos: Video[];
  total: number;
  page: number;
  limit: number;
}

export type VideoWithProgress = Video & {
  progressSeconds?: number;
  progressPercent?: number;
};

export type VideoHistoryEntry = Video & {
  progressSeconds: number;
  progressPercent: number;
  completed: boolean;
  /** When the video was last watched */
  watchedAt: string;
};

export interface WatchHistoryList {
  items: VideoHistoryEntry[];
  total: number;
  page: number;
  limit: number;
}

export interface VideoInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  videoUrl: string;
  thumbnailUrl?: string;
  duration: number;
  genre?: string;
  isPublic?: boolean;
  isPremium?: boolean;
  minimumPlan?: string;
}

export interface VideoUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  thumbnailUrl?: string;
  genre?: string;
  isPublic?: boolean;
  isPremium?: boolean;
  minimumPlan?: string;
  featured?: boolean;
}

export interface WatchProgress {
  videoId: number;
  progressSeconds: number;
  progressPercent: number;
  completed?: boolean;
}

export interface WatchProgressInput {
  progressSeconds: number;
  completed?: boolean;
}

export interface WatchProgressSnapshot {
  videoId: number;
  progressSeconds: number;
  progressPercent: number;
  completed: boolean;
  /** When the video was last watched (preserved on restore) */
  watchedAt: string;
}

export interface WatchProgressSnapshotList {
  items: WatchProgressSnapshot[];
}

export interface LikeStatus {
  liked: boolean;
  likeCount: number;
}

export interface Comment {
  id: number;
  videoId: number;
  /** @nullable */
  parentId?: number | null;
  authorName: string;
  /** @nullable */
  authorAvatar?: string | null;
  authorId?: string;
  body: string;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  body: string;
  parentId?: number;
}

export interface CreatorProfile {
  clerkId: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  videoCount: number;
  totalViews: number;
  totalLikes: number;
  followerCount: number;
}

export interface FollowStatus {
  creatorId: string;
  followerCount: number;
  isFollowing: boolean;
}

export interface Genre {
  id: number;
  name: string;
  slug: string;
  videoCount?: number;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  displayName: string;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  stripeCustomerId?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  displayName?: string;
  username?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface WatchlistItem {
  id: number;
  videoId: number;
  addedAt: string;
}

export interface WatchlistInput {
  videoId: number;
}

export interface Plan {
  id: string;
  name: string;
  priceId: string;
  amount: number;
  currency: string;
  interval: string;
  features: string[];
  /** basic, premium, or ultra */
  tier?: string;
}

export interface SubscriptionStatus {
  isSubscribed: boolean;
  /** @nullable */
  plan: string | null;
  /** @nullable */
  status: string | null;
  /** @nullable */
  currentPeriodEnd?: string | null;
  cancelAtPeriodEnd?: boolean;
}

export interface CheckoutInput {
  priceId: string;
}

export interface CheckoutSession {
  url: string;
}

export interface PortalSession {
  url: string;
}

export type PlatformStatsVideosByGenreItem = {
  genre: string;
  count: number;
};

export interface PlatformStats {
  totalVideos: number;
  totalViews: number;
  totalUsers: number;
  videosByGenre: PlatformStatsVideosByGenreItem[];
}

export interface UploadUrlRequest {
  filename: string;
  contentType: string;
  folder?: string;
}

export interface UploadUrlResponse {
  uploadUrl: string;
  objectPath: string;
}

export type ListVideosParams = {
genre?: string;
search?: string;
page?: number;
limit?: number;
};

export type GetTrendingVideosParams = {
limit?: number;
};

export type GetWatchHistoryParams = {
page?: number;
limit?: number;
/**
 * Filter by completion status. Omit for all, true for finished, false for still watching.
 */
completed?: boolean;
};