# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sineas Video Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscription/checkout:
    post:
      operationId: createCheckout
      tags: [subscription]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{filePath}:
    get:
      tags: [storage]
      operationId: getObject
      summary: Serve a private object entity
      parameters:
        - name: filePath
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema: { type: string, format: binary }
        "404":
          description: Not found
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ─── Video ───────────────────────────────────────────────────────────────
...
    CheckoutInput:
      type: object
      required: [priceId]
      properties:
        priceId: { type: string }
    CheckoutSession:
      type: object
      required: [url]
      properties:
        url: { type: string }
...
    UploadUrlRequest:
      type: object
      required: [filename, contentType]
      properties:
        filename: { type: string }
        contentType: { type: string }
        folder: { type: string }
    UploadUrlResponse:
      type: object
      required: [uploadUrl, objectPath]
      properties:
        uploadUrl: { type: string }
        objectPath: { type: string }